resume-parser/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
pypdf
python-docx==1.1.0
python-multipart==0.0.6
pydantic==2.5.0
loguru==0.7.2
ollama
